✨(s6) allow to shutdown jibri and stop the container
With its internal API, jibri provides a way to shutdown (gracefully or
not) its service. But since jibri is managed by S6 supervisor in this
docker image, the jibri service is always restarted. This is the
default behavior of a S6 service.

We want to shut down the entire container when jibri has been shutdown
or gracefully shutdown.

S6 allows to write custom `finish` scripts for each service. They are
called when the service process is stopped. This commit introduces a
custom handler that stops all services when jibri has been shutdown.